refactor: socket call
